@@ -14,43 +14,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.16",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.16",
-    "@angular/compiler": "^7.2.16",
-    "@angular/core": "^7.2.16",
-    "@angular/forms": "^7.2.16",
-    "@angular/material": "~7.3.7",
-    "@angular/platform-browser": "^7.2.16",
-    "@angular/platform-browser-dynamic": "^7.2.16",
-    "@angular/router": "^7.2.16",
+    "@angular/animations": "^10.1.2",
+    "@angular/cdk": "^10.2.1",
+    "@angular/common": "^10.1.2",
+    "@angular/compiler": "^10.1.2",
+    "@angular/core": "^10.1.2",
+    "@angular/forms": "^10.1.2",
+    "@angular/platform-browser": "^10.1.2",
+    "@angular/platform-browser-dynamic": "^10.1.2",
+    "@angular/router": "^10.1.2",
+    "angular2-draggable": "^2.3.2",
     "core-js": "^2.6.11",
     "fast-xml-parser": "^3.17.4",
     "hammerjs": "^2.0.8",
+    "minimist": "^1.2.5",
+    "ng5-slider": "^1.2.4",
     "normalize-scss": "^7.0.1",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.6.3",
     "tslib": "^1.13.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.10",
-    "@angular/cli": "^7.3.10",
-    "@angular/compiler-cli": "^7.2.16",
-    "@angular/language-service": "^7.2.16",
+    "@angular-devkit/build-angular": "^0.1001.2",
+    "@angular/cli": "^10.1.2",
+    "@angular/compiler-cli": "^10.1.2",
+    "@angular/language-service": "^10.1.2",
     "@types/jasmine": "^2.8.17",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "~8.9.4",
     "codelyzer": "~4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
+    "karma": "^5.2.2",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^2.0.6",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.4",
+    "protractor": "^7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.4"
+    "typescript": "^3.9.7"
   }
 }
