@@ -1,9 +1,4 @@
-import {
-  ModuleWithProviders,
-  NgModule,
-  Optional,
-  SkipSelf,
-} from '@angular/core';
+import { ModuleWithProviders, NgModule, Optional, SkipSelf } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
@@ -20,7 +15,6 @@ import nbSystem from '@i18n/bokmal.system.json';
 import nnSystem from '@i18n/nynorsk.system.json';
 import { ButtonModule } from '../uikit/button/button.module';
 import { SharedModule } from '@shared/shared.module';
-import { MaterialModule } from '../material.module';
 
 @NgModule({
   declarations: [
@@ -39,16 +33,14 @@ import { MaterialModule } from '../material.module';
     ContentModule.forRoot(),
     SharedModule.forRoot(),
     ButtonModule,
-    MaterialModule,
   ],
-  exports: [LayoutComponent],
+  exports: [
+    LayoutComponent,
+  ],
 })
 export class CoreModule {
   /* make sure CoreModule is imported only by one NgModule the AppModule */
-  constructor(
-    @Optional() @SkipSelf() parentModule: CoreModule,
-    contentService: ContentService,
-  ) {
+  constructor(@Optional() @SkipSelf() parentModule: CoreModule, contentService: ContentService) {
     if (parentModule) {
       throw new Error('CoreModule is already loaded. Import only in AppModule');
     }
@@ -58,7 +50,7 @@ export class CoreModule {
     contentService.setCtx('nb');
   }
 
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<CoreModule> {
     return {
       ngModule: CoreModule,
     };
