@@ -1,7 +1,8 @@
-import { Input, OnChanges } from '@angular/core';
+import { Input, OnChanges, Directive } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { QuestionBase } from '@features/questions/interfaces/question-base.interface';
 
+@Directive()
 export class QuestionComponentBase<T = QuestionBase> implements OnChanges {
   @Input()
   public readonly question: T & QuestionBase;
